requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
build==1.2.1
cffi==1.17.1
channels==4.2.0
click==8.1.7
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pip-tools==7.4.1
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.3.0
pyproject_hooks==1.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
service-identity==24.2.0
sqlparse==0.5.0
tomli==2.0.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
zope.interface==7.2
